+++ WeatherAPI/IWeatherService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace WeatherAPI
+{
+    public interface IWeatherService
+    {
+        Task<DataClass> GetWeatherData(string cityName, string stateCode, bool isSuccess);
+    }
+
+}
