+++ WeatherAPI/WeatherForecast.cs
@@ -1,9 +1,10 @@
 ﻿namespace WeatherAPI
 {
-    public class WeatherForecast : iWeather
+    public class WeatherForecast
     {
         public double temp { get; set; }
         public int winddegrees { get; set; }
         public string weather { get; set; }
+        public string winddirection { get; set; }
     }
 }
