--- WeatherAPI/iWeather.cs
@@ -1,11 +1,0 @@
-﻿namespace WeatherAPI
-{
-    interface iWeather
-    {
-        static double temp;
-        static int winddegrees;
-        static string winddirection;
-        static string weather;
-    }
-
-}
